LearnEasyPython
Welcome to LearnEasyPython 🐍 

This is an open source and friendly course where you'll learn the principals of Python 🐍 

It is one of the hottest 🔥 , sexiest 💥 and most demanded 💪🏽  programming languages right now... Surely learning it is super easy, because it has a simple syntax 🈂️  and there are plenty of resources out there in the web 🌐 and yet, here it's another free resource so why not take it? It's free!!! 💸 

This course is meant to have 12 chapters (or modules), which will be always updated:

1. Variables.
2. Data Types.
3. Operators.
4. Lists.
5. Dictionaries.
6. Tuples.
7. Conditionals.
8. Loops.
9. Exception Handling.
10. Functions. 
11. Object-Oriented Programming (OOP).
12. Modules.

I hope you find it useful... and if you like it, please ⭐ 

Happy Coding! 👨🏽‍💻